requirements.txt
groq
python-dotenv
langchain
langchain-community
sentence-transformers
chromadb